remove homebrew-fork os/mac/hardware.rb
* remove many methods
* merge the remainder into hardware.rb directly
* make Hardware a module
* remove :dunno type, return nil instead
* redirect STDERR within backticks
